feat(evolutionary_optimizer): refactor optimizer into mixins and add LLM-driven crossover and mutation
- Split EvolutionaryOptimizer into multiple mixins: LlmMixin, MutationMixin, CrossoverMixin, PopulationMixin, EvaluationMixin, HelpersMixin, StyleMixin
- Add LlmMixin for centralized LLM call handling with rate limiting and caching
- Implement MutationMixin with enhanced semantic, structural, and word-level mutations using LLM guidance
- Implement CrossoverMixin with chunk-level, word-level, and LLM-driven crossover strategies
- Implement PopulationMixin for initializing diverse populations with fresh starts and variations via LLM
- Implement EvaluationMixin for prompt evaluation against datasets
- Implement HelpersMixin for task description generation
- Implement StyleMixin for inferring output style from dataset examples
- Centralize prompt templates and system prompts in evolutionary_optimizer/prompts.py
- Remove deprecated methods and inline prompt strings from main optimizer class
- Improve logging and error handling across mixins

This refactor modularizes the evolutionary optimizer codebase, improves maintainability, and enhances prompt optimization capabilities by leveraging LLMs for mutation, crossover, and population initialization.
