+++ sdks/opik_optimizer/src/opik_optimizer/evolutionary_optimizer/evolutionary_optimizer.py
@@ -1,25 +1,18 @@
 import json
 import logging
-import os
 import random
 from typing import Any, Callable, Dict, List, Optional, Set, Tuple, cast, Type
 
 import rapidfuzz.distance.Indel
-import litellm
 import numpy as np
 import opik
 
 # DEAP imports
 from deap import base, tools
 from deap import creator as _creator
-from litellm import exceptions as litellm_exceptions
-from litellm.caching import Cache
-from litellm.types.caching import LiteLLMCacheType
 from opik.api_objects import opik_client, optimization
 from opik.environment import get_tqdm_for_current_environment
-from opik.evaluation.models.litellm import opik_monitor as opik_litellm_monitor
 
-from opik_optimizer import _throttle, task_evaluator
 from opik_optimizer.base_optimizer import BaseOptimizer, OptimizationRound
 from opik_optimizer.optimization_config import chat_prompt, mappers
 from opik_optimizer.optimization_result import OptimizationResult
@@ -27,19 +20,31 @@ from opik_optimizer.optimizable_agent import OptimizableAgent
 
 from .. import utils
 from . import reporting
+from .llm_mixin import LlmMixin
+from .mutation_mixin import MutationMixin
+from .crossover_mixin import CrossoverMixin
+from .population_mixin import PopulationMixin
+from .evaluation_mixin import EvaluationMixin
+from .helpers_mixin import HelpersMixin
+from .style_mixin import StyleMixin
+from . import prompts as evo_prompts
 
 logger = logging.getLogger(__name__)
 tqdm = get_tqdm_for_current_environment()
-_rate_limiter = _throttle.get_rate_limiter_for_current_opik_installation()
-
-# Using disk cache for LLM calls
-disk_cache_dir = os.path.expanduser("~/.litellm_cache")
-litellm.cache = Cache(type=LiteLLMCacheType.DISK, disk_cache_dir=disk_cache_dir)
 
 creator = cast(Any, _creator)  # type: ignore[assignment]
 
 
-class EvolutionaryOptimizer(BaseOptimizer):
+class EvolutionaryOptimizer(
+    LlmMixin,
+    MutationMixin,
+    CrossoverMixin,
+    PopulationMixin,
+    EvaluationMixin,
+    HelpersMixin,
+    StyleMixin,
+    BaseOptimizer,
+):
     """
     The Evolutionary Optimizer can be used to optimize prompts using a 4 stage genetic algorithm
     approach:
@@ -80,19 +85,7 @@ class EvolutionaryOptimizer(BaseOptimizer):
     )
     DEFAULT_MOO_WEIGHTS = (1.0, -1.0)  # (Maximize Score, Minimize Length)
 
-    _INFER_STYLE_SYSTEM_PROMPT = """You are an expert in linguistic analysis and prompt engineering. Your task is to analyze a few input-output examples from a dataset and provide a concise, actionable description of the desired output style. This description will be used to guide other LLMs in generating and refining prompts.
-
-Focus on characteristics like:
-- **Length**: (e.g., single word, short phrase, one sentence, multiple sentences, a paragraph)
-- **Tone**: (e.g., factual, formal, informal, conversational, academic)
-- **Structure**: (e.g., direct answer first, explanation then answer, list, yes/no then explanation)
-- **Content Details**: (e.g., includes only the answer, includes reasoning, provides examples, avoids pleasantries)
-- **Keywords/Phrasing**: Any recurring keywords or phrasing patterns in the outputs.
-
-Provide a single string that summarizes this style. This summary should be directly usable as an instruction for another LLM.
-For example: 'Outputs should be a single, concise proper noun.' OR 'Outputs should be a short paragraph explaining the reasoning, followed by a direct answer, avoiding conversational pleasantries.' OR 'Outputs are typically 1-2 sentences, providing a direct factual answer.'
-Return ONLY this descriptive string, with no preamble or extra formatting.
-"""
+    # Prompt constants moved into prompts.py
 
     def __init__(
         self,
@@ -286,295 +279,24 @@ Return ONLY this descriptive string, with no preamble or extra formatting.
 
         return total_distance / count if count > 0 else 0.0
 
-    def _deap_crossover_chunking_strategy(
-        self, messages_1_str: str, messages_2_str: str
-    ) -> Tuple[str, str]:
-        chunks1 = [
-            chunk.strip() for chunk in messages_1_str.split(".") if chunk.strip()
-        ]
-        chunks2 = [
-            chunk.strip() for chunk in messages_2_str.split(".") if chunk.strip()
-        ]
+    
 
-        # Try chunk-level crossover if both parents have at least 2 chunks
-        if len(chunks1) >= 2 and len(chunks2) >= 2:
-            min_num_chunks = min(len(chunks1), len(chunks2))
-            # Crossover point is between 1 and min_num_chunks - 1
-            # This requires min_num_chunks >= 2, which is already checked.
-            point = random.randint(1, min_num_chunks - 1)
+    
 
-            child1_chunks = chunks1[:point] + chunks2[point:]
-            child2_chunks = chunks2[:point] + chunks1[point:]
+    
 
-            child1_str = ". ".join(child1_chunks) + ("." if child1_chunks else "")
-            child2_str = ". ".join(child2_chunks) + ("." if child2_chunks else "")
+    
 
-            return child1_str, child2_str
-        else:
-            raise ValueError(
-                "Not enough chunks in either prompt for chunk-level crossover"
-            )
+    
 
-    def _deap_crossover_word_level(
-        self, messages_1_str: str, messages_2_str: str
-    ) -> Tuple[str, str]:
-        words1 = messages_1_str.split()
-        words2 = messages_2_str.split()
-
-        # If either prompt is empty (no words), return parents
-        if not words1 or not words2:
-            return messages_1_str, messages_2_str
-
-        min_word_len = min(len(words1), len(words2))
-        # Need at least 2 words in the shorter prompt for a valid crossover point
-        if min_word_len < 2:
-            return messages_1_str, messages_2_str
-
-        # Crossover point for words: 1 to min_word_len - 1
-        point = random.randint(1, min_word_len - 1)
-        child1_words = words1[:point] + words2[point:]
-        child2_words = words2[:point] + words1[point:]
-
-        return " ".join(child1_words), " ".join(child2_words)
-
-    def _deap_crossover(self, ind1: Any, ind2: Any) -> Tuple[Any, Any]:
-        """Enhanced crossover operation that preserves semantic meaning.
-        Attempts chunk-level crossover first, then falls back to word-level.
-        """
-        reporting.display_message(
-            "      Recombining prompts by mixing and matching words and sentences.",
-            verbose=self.verbose,
-        )
-        messages_1_orig: List[Dict[str, str]] = ind1
-        messages_2_orig: List[Dict[str, str]] = ind2
-
-        for i, message_1 in enumerate(messages_1_orig):
-            role: str = message_1["role"]
-            message_1_str: str = message_1["content"]
-
-            # We check that the second message has enough AI messages and the correct role
-            if (len(messages_2_orig) >= i + 1) and (messages_2_orig[i]["role"] == role):
-                message_2 = messages_2_orig[i]
-                message_2_str: str = message_2["content"]
-
-                try:
-                    child1_str, child2_str = self._deap_crossover_chunking_strategy(
-                        message_1_str, message_2_str
-                    )
-                except ValueError:
-                    child1_str, child2_str = self._deap_crossover_word_level(
-                        message_1_str, message_2_str
-                    )
-
-                # Update the message content
-                messages_1_orig[i]["content"] = child1_str
-                messages_2_orig[i]["content"] = child2_str
-            else:
-                # We don't perform any crossover if there are not enough messages or the roles
-                # don't match
-                pass
-
-        return creator.Individual(messages_1_orig), creator.Individual(messages_2_orig)
-
-    def _deap_mutation(
-        self, individual: Any, initial_prompt: chat_prompt.ChatPrompt
-    ) -> Any:
-        """Enhanced mutation operation with multiple strategies."""
-        prompt = chat_prompt.ChatPrompt(messages=individual)
-
-        # Choose mutation strategy based on current diversity
-        diversity = self._calculate_population_diversity()
-
-        # Determine thresholds based on diversity
-        if diversity < self.DEFAULT_DIVERSITY_THRESHOLD:
-            # Low diversity - use more aggressive mutations (higher chance for semantic)
-            semantic_threshold = 0.5
-            structural_threshold = 0.8  # semantic_threshold + 0.3
-        else:
-            # Good diversity - use more conservative mutations (higher chance for word_level)
-            semantic_threshold = 0.4
-            structural_threshold = 0.7  # semantic_threshold + 0.3
-
-        mutation_choice = random.random()
-
-        if mutation_choice > structural_threshold:
-            # This corresponds to the original 'else' (word_level_mutation)
-            mutated_prompt = self._word_level_mutation_prompt(prompt)
-            reporting.display_success(
-                "      Mutation successful, prompt has been edited by randomizing words (word-level mutation).",
-                verbose=self.verbose,
-            )
-            return creator.Individual(mutated_prompt.get_messages())
-        elif mutation_choice > semantic_threshold:
-            # This corresponds to the original 'elif' (structural_mutation)
-            mutated_prompt = self._structural_mutation(prompt)
-            reporting.display_success(
-                "      Mutation successful, prompt has been edited by reordering, combining, or splitting sentences (structural mutation).",
-                verbose=self.verbose,
-            )
-            return creator.Individual(mutated_prompt.get_messages())
-        else:
-            # This corresponds to the original 'if' (semantic_mutation)
-            mutated_prompt = self._semantic_mutation(prompt, initial_prompt)
-            reporting.display_success(
-                "      Mutation successful, prompt has been edited using an LLM (semantic mutation).",
-                verbose=self.verbose,
-            )
-            return creator.Individual(mutated_prompt.get_messages())
-
-    def _semantic_mutation(
-        self, prompt: chat_prompt.ChatPrompt, initial_prompt: chat_prompt.ChatPrompt
-    ) -> chat_prompt.ChatPrompt:
-        """Enhanced semantic mutation with multiple strategies."""
-        current_output_style_guidance = self.output_style_guidance
-        if random.random() < 0.1:
-            return self._radical_innovation_mutation(prompt, initial_prompt)
-
-        try:
-            strategy = random.choice(
-                [
-                    "rephrase",
-                    "simplify",
-                    "elaborate",
-                    "restructure",
-                    "focus",
-                    "increase_complexity_and_detail",
-                ]
-            )
-
-            strategy_prompts = {
-                "rephrase": f"Create a different way to express the same instruction, possibly with a different length or structure, ensuring it still aims for an answer from the target LLM in the style of: '{current_output_style_guidance}'.",
-                "simplify": f"Simplify the instruction while maintaining its core meaning, potentially making it more concise, to elicit an answer in the style of: '{current_output_style_guidance}'.",
-                "elaborate": f"Add more relevant detail and specificity to the instruction, potentially increasing its length, but only if it helps achieve a more accurate answer from the target LLM in the style of: '{current_output_style_guidance}'.",
-                "restructure": f"Change the structure of the instruction (e.g., reorder sentences, combine/split ideas) while keeping its intent, ensuring the new structure strongly guides towards an output in the style of: '{current_output_style_guidance}'.",
-                "focus": f"Emphasize the key aspects of the instruction, perhaps by rephrasing or adding clarifying statements, to better elicit an answer in the style of: '{current_output_style_guidance}'.",
-                "increase_complexity_and_detail": f"Significantly elaborate on this instruction. Add more details, examples, context, or constraints to make it more comprehensive. The goal of this elaboration is to make the prompt itself more detailed, so that it VERY CLEARLY guides the target LLM to produce a highly accurate final answer in the style of: '{current_output_style_guidance}'. The prompt can be long if needed to achieve this output style.",
-            }
-
-            user_prompt_for_semantic_mutation = f"""Given this prompt: '{prompt}'
-Task context: {self._get_task_description_for_llm(initial_prompt)}
-Desired output style from target LLM: '{current_output_style_guidance}'
-Instruction for this modification: {strategy_prompts[strategy]}.
-Return only the modified prompt message list, nothing else. Make sure to return a valid JSON object.
-"""
-            response = self._call_model(
-                messages=[
-                    {
-                        "role": "system",
-                        "content": f"You are a prompt engineering expert. Your goal is to modify prompts to improve their effectiveness in eliciting specific types of answers, particularly matching the style: '{current_output_style_guidance}'. Follow the specific modification instruction provided.",
-                    },
-                    {"role": "user", "content": user_prompt_for_semantic_mutation},
-                ],
-                is_reasoning=True,
-            )
-
-            return chat_prompt.ChatPrompt(messages=utils.json_to_dict(response.strip()))
-        except Exception as e:
-            reporting.display_error(
-                f"      Error in semantic mutation, this is usually a parsing error: {e}",
-                verbose=self.verbose,
-            )
-            return prompt
-
-    def _structural_mutation(
-        self, prompt: chat_prompt.ChatPrompt
-    ) -> chat_prompt.ChatPrompt:
-        """Perform structural mutation (reordering, combining, splitting)."""
-        mutated_messages: List[Dict[str, str]] = []
-
-        for message in prompt.get_messages():
-            content = message["content"]
-            role = message["role"]
-
-            sentences = [s.strip() for s in content.split(".") if s.strip()]
-            if len(sentences) <= 1:
-                mutated_messages.append(
-                    {"role": role, "content": self._word_level_mutation(content)}
-                )
-                continue
-
-            mutation_type = random.random()
-            if mutation_type < 0.3:
-                # Reorder sentences
-                random.shuffle(sentences)
-                mutated_messages.append(
-                    {"role": role, "content": ". ".join(sentences) + "."}
-                )
-                continue
-            elif mutation_type < 0.6:
-                # Combine adjacent sentences
-                if len(sentences) >= 2:
-                    idx = random.randint(0, len(sentences) - 2)
-                    combined = sentences[idx] + " and " + sentences[idx + 1]
-                    sentences[idx : idx + 2] = [combined]
-                    mutated_messages.append(
-                        {"role": role, "content": ". ".join(sentences) + "."}
-                    )
-                    continue
-            else:
-                # Split a sentence
-                idx = random.randint(0, len(sentences) - 1)
-                words = sentences[idx].split()
-                if len(words) > 3:
-                    split_point = random.randint(2, len(words) - 2)
-                    sentences[idx : idx + 1] = [
-                        " ".join(words[:split_point]),
-                        " ".join(words[split_point:]),
-                    ]
-                    mutated_messages.append(
-                        {"role": role, "content": ". ".join(sentences) + "."}
-                    )
-                    continue
-                else:
-                    mutated_messages.append({"role": role, "content": content})
-
-        return chat_prompt.ChatPrompt(messages=mutated_messages)
-
-    def _word_level_mutation_prompt(
-        self, prompt: chat_prompt.ChatPrompt
-    ) -> chat_prompt.ChatPrompt:
-        mutated_messages: List[Dict[str, str]] = []
-        for message in prompt.get_messages():
-            mutated_messages.append(
-                {
-                    "role": message["role"],
-                    "content": self._word_level_mutation(message["content"]),
-                }
-            )
-        return chat_prompt.ChatPrompt(messages=mutated_messages)
-
-    def _word_level_mutation(self, msg_content: str) -> str:
-        """Perform word-level mutation."""
-        words = msg_content.split()
-        if len(words) <= 1:
-            return msg_content
-
-        mutation_type = random.random()
-        if mutation_type < 0.3:
-            # Word replacement
-            idx = random.randint(0, len(words) - 1)
-            words[idx] = self._get_synonym(words[idx])
-        elif mutation_type < 0.6:
-            # Word reordering
-            if len(words) > 2:
-                i, j = random.sample(range(len(words)), 2)
-                words[i], words[j] = words[j], words[i]
-        else:
-            # Phrase modification
-            idx = random.randint(0, len(words) - 1)
-            words[idx] = self._modify_phrase(words[idx])
-
-        return " ".join(words)
+    
 
     def _get_synonym(self, word: str) -> str:
         """Get a synonym for a word using LLM."""
         try:
             response = self._call_model(
                 messages=[
-                    {
-                        "role": "system",
-                        "content": "You are a helpful assistant that provides synonyms. Return only the synonym word, no explanation or additional text.",
-                    },
+                    {"role": "system", "content": evo_prompts.synonyms_system_prompt()},
                     {
                         "role": "user",
                         "content": f"Give me a single synonym for the word '{word}'. Return only the synonym, nothing else.",
@@ -592,10 +314,7 @@ Return only the modified prompt message list, nothing else. Make sure to return 
         try:
             response = self._call_model(
                 messages=[
-                    {
-                        "role": "system",
-                        "content": "You are a helpful assistant that rephrases text. Return only the modified phrase, no explanation or additional text.",
-                    },
+                    {"role": "system", "content": evo_prompts.rephrase_system_prompt()},
                     {
                         "role": "user",
                         "content": f"Modify this phrase while keeping the same meaning: '{phrase}'. Return only the modified phrase, nothing else.",
@@ -634,7 +353,9 @@ Return only the new prompt list object.
                 messages=[
                     {
                         "role": "system",
-                        "content": self._get_radical_innovation_system_prompt(),
+                        "content": evo_prompts.radical_innovation_system_prompt(
+                            self.output_style_guidance
+                        ),
                     },
                     {"role": "user", "content": user_prompt_for_radical_innovation},
                 ],
@@ -650,188 +371,7 @@ Return only the new prompt list object.
             )
             return prompt
 
-    def _initialize_population(
-        self, prompt: chat_prompt.ChatPrompt
-    ) -> List[chat_prompt.ChatPrompt]:
-        """Initialize the population with diverse variations of the initial prompt,
-        including some 'fresh start' prompts based purely on task description.
-        All generated prompts should aim to elicit answers matching self.output_style_guidance.
-        """
-        with reporting.initializing_population(verbose=self.verbose) as init_pop_report:
-            init_pop_report.start(self.population_size)
-
-            population = [prompt]
-            if self.population_size <= 1:
-                return population
-
-            num_to_generate_total = self.population_size - 1
-            num_fresh_starts = max(1, int(num_to_generate_total * 0.2))
-            num_variations_on_initial = num_to_generate_total - num_fresh_starts
-
-            task_desc_for_llm = self._get_task_description_for_llm(prompt)
-            current_output_style_guidance = self.output_style_guidance
-
-            # Generate "fresh start" prompts if the initial prompt is not performing well
-            # Cold start prompts are generated from the task description
-            if num_fresh_starts > 0:
-                init_pop_report.start_fresh_prompts(num_fresh_starts)
-                fresh_start_user_prompt = f"""Here is a description of a task:
-    {task_desc_for_llm}
-
-    The goal is to generate prompts that will make a target LLM produce responses in the following style: '{current_output_style_guidance}'.
-
-    Please generate {num_fresh_starts} diverse and effective prompt(s) for a language model to accomplish this task, ensuring they guide towards this specific output style.
-    Focus on clarity, completeness, and guiding the model effectively towards the desired style. Explore different structural approaches.
-
-    Example of valid response: [
-        ["role": "<role>", "content": "<Prompt targeting specified style.>"],
-        ["role": "<role>", "content": "<Another prompt designed for the output style.>"]
-    ]
-
-    Your response MUST be a valid JSON list of AI messages. Do NOT include any other text, explanations, or Markdown formatting like ```json ... ``` around the list.
-
-    """
-                try:
-                    response_content = self._call_model(
-                        messages=[
-                            {
-                                "role": "system",
-                                "content": f"You are an expert prompt engineer. Your task is to generate novel, effective prompts from scratch based on a task description, specifically aiming for prompts that elicit answers in the style: '{current_output_style_guidance}'. Output ONLY a raw JSON list of strings.",
-                            },
-                            {"role": "user", "content": fresh_start_user_prompt},
-                        ],
-                        is_reasoning=True,
-                    )
-
-                    logger.debug(
-                        f"Raw LLM response for fresh start prompts: {response_content}"
-                    )
-
-                    fresh_prompts = utils.json_to_dict(response_content)
-                    if isinstance(fresh_prompts, list):
-                        if all(isinstance(p, dict) for p in fresh_prompts) and all(
-                            p.get("role") is not None for p in fresh_prompts
-                        ):
-                            population.append(
-                                chat_prompt.ChatPrompt(messages=fresh_prompts)
-                            )
-                            init_pop_report.success_fresh_prompts(1)
-                        elif all(isinstance(p, list) for p in fresh_prompts):
-                            population.extend(
-                                [
-                                    chat_prompt.ChatPrompt(messages=p)
-                                    for p in fresh_prompts[:num_fresh_starts]
-                                ]
-                            )
-                            init_pop_report.success_fresh_prompts(
-                                len(fresh_prompts[:num_fresh_starts])
-                            )
-                        else:
-                            init_pop_report.failed_fresh_prompts(
-                                num_fresh_starts,
-                                f"LLM response for fresh starts was not a valid list of strings or was empty: {response_content}. Skipping fresh start prompts.",
-                            )
-                except json.JSONDecodeError as e_json:
-                    init_pop_report.failed_fresh_prompts(
-                        num_fresh_starts,
-                        f"JSONDecodeError generating fresh start prompts: {e_json}. LLM response: '{response_content}'. Skipping fresh start prompts.",
-                    )
-                except Exception as e:
-                    init_pop_report.failed_fresh_prompts(
-                        num_fresh_starts,
-                        f"Error generating fresh start prompts: {e}. Skipping fresh start prompts.",
-                    )
-
-            # Generate variations on the initial prompt for the remaining slots
-            # TODO: Could add variations with hyper-parameters from the task config like temperature, etc.
-            if num_variations_on_initial > 0:
-                init_pop_report.start_variations(num_variations_on_initial)
-
-                # TODO: We need to split this into batches as the model will not return enough tokens
-                # to generate all the candidates
-                user_prompt_for_variation = f"""Initial prompt:
-    '''{prompt.get_messages()}'''
-
-    Task context:
-    {task_desc_for_llm}
-    Desired output style from target LLM: '{current_output_style_guidance}'
-
-    Generate {num_variations_on_initial} diverse alternative prompts based on the initial prompt above, keeping the task context and desired output style in mind.
-    All generated prompt variations should strongly aim to elicit answers from the target LLM matching the style: '{current_output_style_guidance}'.
-    For each variation, consider how to best achieve this style, e.g., by adjusting specificity, structure, phrasing, constraints, or by explicitly requesting it.
-
-    Return a JSON array of prompts with the following structure:
-    {{
-        "prompts": [
-            {{
-                "prompt": [{{"role": "<role>", "content": "<content>"}}],
-                "strategy": "brief description of the variation strategy used, e.g., 'direct instruction for target style'"
-            }}
-            // ... more prompts if num_variations_on_initial > 1
-        ]
-    }}
-    Ensure a good mix of variations, all targeting the specified output style from the end LLM.
-
-    Return a valid JSON object that is correctly escaped. Return nothing else, d`o not include any additional text or Markdown formatting.
-    """
-                try:
-                    response_content_variations = self._call_model(
-                        messages=[
-                            {
-                                "role": "system",
-                                "content": self._get_reasoning_system_prompt_for_variation(),
-                            },
-                            {"role": "user", "content": user_prompt_for_variation},
-                        ],
-                        is_reasoning=True,
-                    )
-                    logger.debug(
-                        f"Raw response for population variations: {response_content_variations}"
-                    )
-                    json_response_variations = json.loads(response_content_variations)
-                    generated_prompts_variations = [
-                        p["prompt"]
-                        for p in json_response_variations.get("prompts", [])
-                        if isinstance(p, dict) and "prompt" in p
-                    ]
-
-                    if generated_prompts_variations:
-                        init_pop_report.success_variations(
-                            len(
-                                generated_prompts_variations[:num_variations_on_initial]
-                            )
-                        )
-                        population.extend(
-                            [
-                                chat_prompt.ChatPrompt(messages=p)
-                                for p in generated_prompts_variations[
-                                    :num_variations_on_initial
-                                ]
-                            ]
-                        )
-                    else:
-                        init_pop_report.failed_variations(
-                            num_variations_on_initial,
-                            "Could not parse 'prompts' list for variations. Skipping variations.",
-                        )
-                except Exception as e:
-                    init_pop_report.failed_variations(
-                        num_variations_on_initial,
-                        f"Error calling LLM for initial population variations: {e}",
-                    )
-
-            # Ensure population is of the required size using unique prompts
-            # TODO Test with levenshtein distance
-            final_population_set: Set[str] = set()
-            final_population_list: List[chat_prompt.ChatPrompt] = []
-            for p in population:
-                if json.dumps(p.get_messages()) not in final_population_set:
-                    final_population_set.add(json.dumps(p.get_messages()))
-                    final_population_list.append(p)
-
-            init_pop_report.end(final_population_list)
-            # Return exactly population_size prompts if possible, or fewer if generation failed badly.
-            return final_population_list[: self.population_size]
+    
 
     def _should_restart_population(self, curr_best: float) -> bool:
         """
@@ -1434,73 +974,6 @@ Return only the new prompt list object.
             optimization_id=self._current_optimization_id,
         )
 
-    @_throttle.rate_limited(_rate_limiter)
-    def _call_model(
-        self,
-        messages: List[Dict[str, str]],
-        is_reasoning: bool = False,
-        optimization_id: Optional[str] = None,
-    ) -> str:
-        """Call the model with the given prompt and return the response."""
-        try:
-            # Basic LLM parameters
-            llm_config_params = {
-                "temperature": getattr(self, "temperature", 0.3),
-                "max_tokens": getattr(self, "max_tokens", 1000),
-                "top_p": getattr(self, "top_p", 1.0),
-                "frequency_penalty": getattr(self, "frequency_penalty", 0.0),
-                "presence_penalty": getattr(self, "presence_penalty", 0.0),
-            }
-
-            # Prepare metadata for opik
-            metadata_for_opik: Dict[str, Any] = {}
-            if self.project_name:
-                metadata_for_opik["project_name"] = self.project_name
-                metadata_for_opik["opik"] = {"project_name": self.project_name}
-
-            if optimization_id:
-                if "opik" in metadata_for_opik:
-                    metadata_for_opik["opik"]["optimization_id"] = optimization_id
-
-            metadata_for_opik["optimizer_name"] = self.__class__.__name__
-            metadata_for_opik["opik_call_type"] = (
-                "reasoning" if is_reasoning else "evaluation_llm_task_direct"
-            )
-
-            if metadata_for_opik:
-                llm_config_params["metadata"] = metadata_for_opik
-
-            # Pass llm_config_params to the Opik monitor
-            final_call_params = opik_litellm_monitor.try_add_opik_monitoring_to_params(
-                llm_config_params.copy()
-            )
-
-            logger.debug(
-                f"Calling model '{self.model}' with messages: {messages}, "
-                f"final params for litellm (from monitor): {final_call_params}"
-            )
-
-            response = litellm.completion(
-                model=self.model, messages=messages, **final_call_params
-            )
-            self.llm_call_counter += 1
-
-            logger.debug(f"Response: {response}")
-            return response.choices[0].message.content
-        except litellm_exceptions.RateLimitError as e:
-            logger.error(f"LiteLLM Rate Limit Error: {e}")
-            raise
-        except litellm_exceptions.APIConnectionError as e:
-            logger.error(f"LiteLLM API Connection Error: {e}")
-            raise
-        except litellm_exceptions.ContextWindowExceededError as e:
-            logger.error(f"LiteLLM Context Window Exceeded Error: {e}")
-            raise
-        except Exception as e:
-            logger.error(
-                f"Error calling model '{self.model}': {type(e).__name__} - {e}"
-            )
-            raise
 
     def _evaluate_prompt(
         self,
@@ -1613,7 +1086,9 @@ Follow the instructions provided in the system prompt regarding the JSON output 
                 messages=[
                     {
                         "role": "system",
-                        "content": self.get_llm_crossover_system_prompt(),
+                        "content": evo_prompts.llm_crossover_system_prompt(
+                            current_output_style_guidance
+                        ),
                     },
                     {"role": "user", "content": user_prompt_for_llm_crossover},
                 ],
@@ -1654,35 +1129,7 @@ Follow the instructions provided in the system prompt regarding the JSON output 
         return description
 
     def _get_reasoning_system_prompt_for_variation(self) -> str:
-        return f"""You are an expert prompt engineer specializing in creating diverse and effective prompts. Given an initial prompt, your task is to generate a diverse set of alternative prompts.
-
-For each prompt variation, consider:
-1. Different levels of specificity and detail, including significantly more detailed and longer versions.
-2. Various ways to structure the instruction, exploring more complex sentence structures and phrasings.
-3. Alternative phrasings that maintain the core intent but vary in style and complexity.
-4. Different emphasis on key components, potentially elaborating on them.
-5. Various ways to express constraints or requirements.
-6. Different approaches to clarity and conciseness, but also explore more verbose and explanatory styles.
-7. Alternative ways to guide the model's response format.
-8. Consider variations that are substantially longer and more descriptive than the original.
-
-The generated prompts should guide a target LLM to produce outputs in the following style: '{self.output_style_guidance}'
-
-Return a JSON array of prompts with the following structure:
-{{
-    "prompts": [
-        {{
-            "prompt": "alternative prompt 1",
-            "strategy": "brief description of the variation strategy used, e.g., 'focused on eliciting specific output style'"
-        }},
-        {{
-            "prompt": "alternative prompt 2",
-            "strategy": "brief description of the variation strategy used"
-        }}
-    ]
-}}
-Each prompt variation should aim to get the target LLM to produce answers matching the desired style: '{self.output_style_guidance}'.
-"""
+        return evo_prompts.variation_system_prompt(self.output_style_guidance)
 
     def get_llm_crossover_system_prompt(self) -> str:
         return f"""You are an expert prompt engineer specializing in creating novel prompts by intelligently blending existing ones.
@@ -1707,12 +1154,7 @@ Return a JSON object that is a list of both child prompts. Each child prompt is 
 """
 
     def _get_radical_innovation_system_prompt(self) -> str:
-        return f"""You are an expert prompt engineer and a creative problem solver.
-Given a task description and an existing prompt for that task (which might be underperforming), your goal is to generate a new, significantly improved, and potentially very different prompt.
-Do not just make minor edits. Think about alternative approaches, structures, and phrasings that could lead to better performance.
-Consider clarity, specificity, constraints, and how to best guide the language model for the described task TO PRODUCE OUTPUTS IN THE FOLLOWING STYLE: '{self.output_style_guidance}'.
-Return only the new prompt string, with no preamble or explanation.
-"""
+        return evo_prompts.radical_innovation_system_prompt(self.output_style_guidance)
 
     def _infer_output_style_from_dataset(
         self, dataset: opik.Dataset, prompt: chat_prompt.ChatPrompt, n_examples: int = 5
@@ -1730,6 +1172,16 @@ Return only the new prompt string, with no preamble or explanation.
                     f"Failed to get items from dataset '{dataset.name}': {e}"
                 )
                 return None
+
+    # Override prompt builders to centralize strings in prompts.py
+    def _get_reasoning_system_prompt_for_variation(self) -> str:  # type: ignore[override]
+        return evo_prompts.variation_system_prompt(self.output_style_guidance)
+
+    def get_llm_crossover_system_prompt(self) -> str:  # type: ignore[override]
+        return evo_prompts.llm_crossover_system_prompt(self.output_style_guidance)
+
+    def _get_radical_innovation_system_prompt(self) -> str:  # type: ignore[override]
+        return evo_prompts.radical_innovation_system_prompt(self.output_style_guidance)
 
             if not items_to_process:
                 report_infer_output_style.error(
@@ -1765,7 +1217,7 @@ Return only the new prompt string, with no preamble or explanation.
             try:
                 inferred_style = self._call_model(
                     messages=[
-                        {"role": "system", "content": self._INFER_STYLE_SYSTEM_PROMPT},
+                        {"role": "system", "content": evo_prompts.INFER_STYLE_SYSTEM_PROMPT},
                         {"role": "user", "content": user_prompt_for_style_inference},
                     ],
                     is_reasoning=True,
